changed snt atom type from float to double

+++ src/snt/exs/atom.h
@@ -57,7 +57,7 @@ namespace snt::exs {
     }
   };
 
-  typedef std::variant<float, bool> AtomValueType;
+  typedef std::variant<double, bool> AtomValueType;
 
   class Atom : public AtomBase<AtomValueType> {
   public:
@@ -82,9 +82,9 @@ namespace snt::exs {
       return v;
     }
     std::string to_string() {
-      if (std::holds_alternative<float>(value)) {
+      if (std::holds_alternative<double>(value)) {
         std::stringstream str;
-        str << std::get<float>(value) << std::scientific;
+        str << std::get<double>(value) << std::scientific;
         return str.str();
       } else {
         if (std::get<bool>(value) == 0)
@@ -95,77 +95,77 @@ namespace snt::exs {
     }
     // Math operations
     void math_add(Atom* other) {
-      value = std::get<float>(value) + std::get<float>(other->value);
+      value = std::get<double>(value) + std::get<double>(other->value);
     }
     void math_subtract(Atom* other) {
-      value = std::get<float>(value) - std::get<float>(other->value);
+      value = std::get<double>(value) - std::get<double>(other->value);
     }
     void math_negate() {
-      value = -std::get<float>(value);
+      value = -std::get<double>(value);
     }
     void math_multiply(Atom* other) {
-      value = std::get<float>(value) * std::get<float>(other->value);
+      value = std::get<double>(value) * std::get<double>(other->value);
     }
     void math_divide(Atom* other) {
-      value = std::get<float>(value) / std::get<float>(other->value);
+      value = std::get<double>(value) / std::get<double>(other->value);
     }
     void math_power(Atom* other) {
-      value = pow(std::get<float>(value), std::get<float>(other->value));
+      value = pow(std::get<double>(value), std::get<double>(other->value));
     }
     void math_modulo(Atom* other) {
-      value = (float)((int)std::get<float>(value) % (int)std::get<float>(other->value));
+      value = (double)((int)std::get<double>(value) % (int)std::get<double>(other->value));
     }
 
     // Argument math operators
     void math_exponent() {
-      value = std::exp(std::get<float>(value));
+      value = std::exp(std::get<double>(value));
     }
     void math_logarithm() {
-      value = std::log(std::get<float>(value));
+      value = std::log(std::get<double>(value));
     }
     void math_logarithm_10() {
-      value = std::log10(std::get<float>(value));
+      value = std::log10(std::get<double>(value));
     }
     void math_logarithm_base(Atom* other) {
-      value = std::log(std::get<float>(value)) / std::log(std::get<float>(other->value));
+      value = std::log(std::get<double>(value)) / std::log(std::get<double>(other->value));
     }
     void math_power_base(Atom* other) {
-      value = pow(std::get<float>(value), std::get<float>(other->value));
+      value = pow(std::get<double>(value), std::get<double>(other->value));
     }
     void math_square_root() {
-      value = std::sqrt(std::get<float>(value));
+      value = std::sqrt(std::get<double>(value));
     }
     void math_cubic_root() {
-      value = std::cbrt(std::get<float>(value));
+      value = std::cbrt(std::get<double>(value));
     }
     void math_sinus() {
-      value = std::sin(std::get<float>(value));
+      value = std::sin(std::get<double>(value));
     }
     void math_cosinus() {
-      value = std::cos(std::get<float>(value));
+      value = std::cos(std::get<double>(value));
     }
     void math_tangens() {
-      value = std::tan(std::get<float>(value));
+      value = std::tan(std::get<double>(value));
     }
 
     // Comparison operations
     void comparison_equal(Atom* other) {
-      value = std::get<float>(value) == std::get<float>(other->value);
+      value = std::get<double>(value) == std::get<double>(other->value);
     }
     void comparison_not_equal(Atom* other) {
-      value = std::get<float>(value) != std::get<float>(other->value);
+      value = std::get<double>(value) != std::get<double>(other->value);
     }
     void comparison_less_equal(Atom* other) {
-      value = std::get<float>(value) <= std::get<float>(other->value);
+      value = std::get<double>(value) <= std::get<double>(other->value);
     }
     void comparison_greater_equal(Atom* other) {
-      value = std::get<float>(value) >= std::get<float>(other->value);
+      value = std::get<double>(value) >= std::get<double>(other->value);
     }
     void comparison_less(Atom* other) {
-      value = std::get<float>(value) < std::get<float>(other->value);
+      value = std::get<double>(value) < std::get<double>(other->value);
     }
     void comparison_greater(Atom* other) {
-      value = std::get<float>(value) > std::get<float>(other->value);
+      value = std::get<double>(value) > std::get<double>(other->value);
     }
 
     // Logical operations
@@ -181,7 +181,7 @@ namespace snt::exs {
 
     // Conditional operator
     void condition(Atom* option1, Atom* option2) {
-      value = std::get<bool>(value) ? std::get<float>(option1->value) : std::get<float>(option2->value);
+      value = std::get<bool>(value) ? std::get<double>(option1->value) : std::get<double>(option2->value);
     }
   };
